requirements.txt
Authlib==0.11
Flask==1.0.2
cryptography==2.6.1
google-cloud-datastore==1.7.3
google-cloud-kms==1.0.0
google-cloud-logging==1.10.0
itsdangerous==1.1.0
requests==2.21.0
websockets==7.0
Werkzeug==0.15.2
google-auth-oauthlib==0.3.0
pytest==4.5.0